refactor games per platform
